Fix OS.execute_with_pipe usage
The problem seems to be with the get_as_text method which seems to want
to read all the content at once which doesn't work for stdout.
We can instead use get_line, get_int, get_float etc, i.e functions which
try to read a limited amount of data.

This commit also restructures the code to launch the process in a
seperate thread and doesn't launch the process through cmd.exe on
Windows.